add docblocks and cs fix #ITR-95
