+++ src/CustomExceptions/actionTagException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class actionTagException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class actionTagException extends Exception
+{
+}
