+++ src/CustomExceptions/authException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class authException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class authException extends Exception
+{
+}
