+++ src/CustomExceptions/configurationException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class configurationException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class configurationException extends Exception
+{
+}
