+++ src/CustomExceptions/countryException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class countryException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class countryException extends Exception
+{
+}
