+++ src/CustomExceptions/credentialsException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class credentialsException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class credentialsException extends Exception
+{
+}
