+++ src/CustomExceptions/generalException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class generalException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class generalException extends Exception
+{
+}
