+++ src/CustomExceptions/htmlTagException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class htmlTagException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class htmlTagException extends Exception
+{
+}
