+++ src/CustomExceptions/languageException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class languageException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class languageException extends Exception
+{
+}
