+++ src/CustomExceptions/notSimpleXmlInstanceException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class notSimpleXmlInstanceException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class notSimpleXmlInstanceException extends Exception
+{
+}
