+++ src/CustomExceptions/pdfFilenameException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class pdfFilenameException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class pdfFilenameException extends Exception
+{
+}
