+++ src/CustomExceptions/pdfMD5Exception.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class pdfMD5Exception
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class pdfMD5Exception extends Exception
+{
+}
