+++ src/CustomExceptions/pdfUrlException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class pdfUrlException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class pdfUrlException extends Exception
+{
+}
