+++ src/CustomExceptions/postPageException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class postPageException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class postPageException extends Exception
+{
+}
