+++ src/CustomExceptions/textException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class textException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class textException extends Exception
+{
+}
