+++ src/CustomExceptions/textTypeException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class textTypeException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class textTypeException extends Exception
+{
+}
