+++ src/CustomExceptions/titleException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class titleException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class titleException extends Exception
+{
+}
