+++ src/CustomExceptions/versionException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Inpsyde\AGBConnector\CustomExceptions;
+
+use Exception;
+
+/**
+ * Class versionException
+ *
+ * @package Inpsyde\AGBConnector\CustomExceptions
+ */
+class versionException extends Exception
+{
+}
