+++ src/Middleware/CheckVersionXml.php
@@ -1,23 +1,27 @@
 <?php
 
-
 namespace Inpsyde\AGBConnector\Middleware;
 
-
-use Inpsyde\AGBConnector\customExceptions\versionException;
+use Exception;
+use Inpsyde\AGBConnector\CustomExceptions\versionException;
 use Inpsyde\AGBConnector\XmlApi;
 
+/**
+ * Class CheckVersionXml
+ *
+ * @package Inpsyde\AGBConnector\Middleware
+ */
 class CheckVersionXml extends Middleware
 {
+
     /**
      * @param $xml
      *
-     * @return bool|\Exception|versionException\
+     * @return Exception|versionException|int
      */
     public function process($xml)
     {
         try {
-
             if (XmlApi::VERSION !== (string)$xml->api_version) {
                 throw new versionException(
                     "Version Exception: provided {$xml->api_version}",
