--- src/customExceptions/actionTagException.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Inpsyde\AGBConnector\customExceptions;
-
-
-use Exception;
-
-class actionTagException extends Exception
-{
-}
