--- src/customExceptions/configurationException.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Inpsyde\AGBConnector\customExceptions;
-
-
-use Exception;
-
-class configurationException extends Exception
-{
-}
