--- src/customExceptions/pdfFilenameException.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Inpsyde\AGBConnector\customExceptions;
-
-
-use Exception;
-
-class pdfFilenameException extends Exception
-{
-}
