--- src/customExceptions/pdfMD5Exception.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Inpsyde\AGBConnector\customExceptions;
-
-
-use Exception;
-
-class pdfMD5Exception extends Exception
-{
-}
