--- src/customExceptions/textTypeException.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace Inpsyde\AGBConnector\customExceptions;
-
-
-use Exception;
-
-class textTypeException extends Exception
-{
-}
